// src/api/models/ProjectRead.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ProjectRead = {
    title: string;
    description: string;
    categories: Array<string>;
    lorem1: string;
    lorem2: string;
    lorem3: number;
    lorem4: number;
    lorem5: string;
    lorem6: string;
    lorem7: boolean;
    lorem8: string;
    lorem9: Array<string>;
    lorem10: string;
    id: number;
};